fix conflict with base_transformer PR
